client: Move CNI store into client package.
The CNI store is specific to the client, so it should be in that
package not a more generic location.
